Separated engine and move code from board.h and implemented a basic engine constructor
