+++ cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/DependInfo.cmake
@@ -5,6 +5,7 @@ set(CMAKE_DEPENDS_LANGUAGES
 # The set of files for implicit dependencies of each language:
 set(CMAKE_DEPENDS_CHECK_CXX
   "/home/sedano/CLionProjects/MakeDamaGreatAgain/board/board.cpp" "/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/board/board.cpp.o"
+  "/home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.cpp" "/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o"
   "/home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.cpp" "/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/helpers/helpers.cpp.o"
   "/home/sedano/CLionProjects/MakeDamaGreatAgain/main.cpp" "/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/main.cpp.o"
   "/home/sedano/CLionProjects/MakeDamaGreatAgain/renders/render.cpp" "/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.o"
