+++ cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/build.make
@@ -128,12 +128,26 @@ CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.s"
 	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/sedano/CLionProjects/MakeDamaGreatAgain/renders/render.cpp -o CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.s
 
+CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o: CMakeFiles/MakeDamaGreatAgain.dir/flags.make
+CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o: ../engine/engine.cpp
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_5) "Building CXX object CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o"
+	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o -c /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.cpp
+
+CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.i"
+	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.cpp > CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.i
+
+CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.s"
+	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.cpp -o CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.s
+
 # Object files for target MakeDamaGreatAgain
 MakeDamaGreatAgain_OBJECTS = \
 "CMakeFiles/MakeDamaGreatAgain.dir/main.cpp.o" \
 "CMakeFiles/MakeDamaGreatAgain.dir/board/board.cpp.o" \
 "CMakeFiles/MakeDamaGreatAgain.dir/helpers/helpers.cpp.o" \
-"CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.o"
+"CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.o" \
+"CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o"
 
 # External object files for target MakeDamaGreatAgain
 MakeDamaGreatAgain_EXTERNAL_OBJECTS =
@@ -142,9 +156,10 @@ MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/main.cpp.o
 MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/board/board.cpp.o
 MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/helpers/helpers.cpp.o
 MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.o
+MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o
 MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/build.make
 MakeDamaGreatAgain: CMakeFiles/MakeDamaGreatAgain.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_5) "Linking CXX executable MakeDamaGreatAgain"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_6) "Linking CXX executable MakeDamaGreatAgain"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/MakeDamaGreatAgain.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
