+++ cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/cmake_clean.cmake
@@ -1,5 +1,6 @@
 file(REMOVE_RECURSE
   "CMakeFiles/MakeDamaGreatAgain.dir/board/board.cpp.o"
+  "CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o"
   "CMakeFiles/MakeDamaGreatAgain.dir/helpers/helpers.cpp.o"
   "CMakeFiles/MakeDamaGreatAgain.dir/main.cpp.o"
   "CMakeFiles/MakeDamaGreatAgain.dir/renders/render.cpp.o"
