+++ cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/depend.internal
@@ -5,11 +5,17 @@ CMakeFiles/MakeDamaGreatAgain.dir/board/board.cpp.o
  /home/sedano/CLionProjects/MakeDamaGreatAgain/board/board.cpp
  /home/sedano/CLionProjects/MakeDamaGreatAgain/board/board.h
  /home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.h
+CMakeFiles/MakeDamaGreatAgain.dir/engine/engine.cpp.o
+ /home/sedano/CLionProjects/MakeDamaGreatAgain/board/board.h
+ /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.cpp
+ /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.h
+ /home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.h
 CMakeFiles/MakeDamaGreatAgain.dir/helpers/helpers.cpp.o
  /home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.cpp
  /home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.h
 CMakeFiles/MakeDamaGreatAgain.dir/main.cpp.o
  /home/sedano/CLionProjects/MakeDamaGreatAgain/board/board.h
+ /home/sedano/CLionProjects/MakeDamaGreatAgain/engine/engine.h
  /home/sedano/CLionProjects/MakeDamaGreatAgain/helpers/helpers.h
  /home/sedano/CLionProjects/MakeDamaGreatAgain/main.cpp
  /home/sedano/CLionProjects/MakeDamaGreatAgain/renders/render.h
