+++ cmake-build-debug/CMakeFiles/MakeDamaGreatAgain.dir/progress.make
@@ -3,4 +3,5 @@ CMAKE_PROGRESS_2 = 2
 CMAKE_PROGRESS_3 = 3
 CMAKE_PROGRESS_4 = 4
 CMAKE_PROGRESS_5 = 5
+CMAKE_PROGRESS_6 = 6
