+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -93,12 +93,12 @@ clean: CMakeFiles/MakeDamaGreatAgain.dir/clean
 CMakeFiles/MakeDamaGreatAgain.dir/all:
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/MakeDamaGreatAgain.dir/build.make CMakeFiles/MakeDamaGreatAgain.dir/depend
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/MakeDamaGreatAgain.dir/build.make CMakeFiles/MakeDamaGreatAgain.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4,5 "Built target MakeDamaGreatAgain"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4,5,6 "Built target MakeDamaGreatAgain"
 .PHONY : CMakeFiles/MakeDamaGreatAgain.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/MakeDamaGreatAgain.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles 5
+	$(CMAKE_COMMAND) -E cmake_progress_start /home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles 6
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/Makefile2 CMakeFiles/MakeDamaGreatAgain.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start /home/sedano/CLionProjects/MakeDamaGreatAgain/cmake-build-debug/CMakeFiles 0
 .PHONY : CMakeFiles/MakeDamaGreatAgain.dir/rule
