+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Jun 26 17:45 CEST
+Start testing: Jun 26 18:19 CEST
 ----------------------------------------------------------
-End testing: Jun 26 17:45 CEST
+End testing: Jun 26 18:19 CEST
